+++ lib/amino_acid/amino_acid_all.dart
@@ -5,3 +5,5 @@ export 'enum_amino_acid.dart';
 export 'enum_amino_acid_sequence_direction.dart';
 export 'f_amino_acid.dart';
 export 'util_amino_acid.dart';
+export 'util_amino_acid_search.dart';
+export 'util_compare_amino_acid.dart';