+++ lib/converter/converter_all.dart
@@ -1,2 +1,3 @@
 export 'fasta/fasta_record.dart';
 export 'fasta/util_fasta.dart';
+export 'fasta/util_fasta_search.dart';